+++ CHANGES.rst
@@ -81,6 +81,11 @@ tweakreg
 - Fix a bug due to which source catalog may contain sources
   outside of the bounding box. [#947]
 
+source_detection
+----------------
+
+- Support for PSF fitting (optional) for accurate centroids. [#841]
+
 0.12.0 (2023-08-18)
 ===================